fix[migrations]: Resolved an error with migrations in the blog application
An error occurred when adding a non-nullable field 'release_year' to the 'blogpost'
model without specifying a default. The database needed a value to populate existing rows.
When the migrations were run, it resulted in a 'DuplicateColumn' error as the column 'release_year'
already existed. This fix resolves the issue by adding a default value to the 'release_year' field to prevent the error during migrations.

On branch main
Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   blog/admin.py
	modified:   blog/apps.py
	new file:   blog/migrations/0002_blogpost_release_year_alter_blogpost_content_and_more.py
	modified:   blog/models.py